+++ src/Redirect.tsx
@@ -51,8 +51,12 @@ function Redirect(props: RedirectProps) {
 						onMount={() => {
 							method(location);
 						}}
-						onUpdate={(self, prevProps) => {
-							if (!locationsAreEqual(prevProps.to, location)) {
+						onUpdate={(self, prevProps: RedirectProps) => {
+							if (typeof prevProps.to === "string") {
+								if (prevProps.to !== location.pathname) {
+									method(location);
+								}
+							} else if (!locationsAreEqual(prevProps.to, location)) {
 								method(location);
 							}
 						}}